Table writer 10: Support partition write in TableWriter
When the table is partitioned, create multiple dwrf writers
by HiveDataSink, one for each distinct partition value.
Call PartitionIdGenerator to generate consecutive IDs
for distinct partition values, which are used for the index
of the dwrf writer vector. Each input could see new partition
values, so dwrf writers are created incrementally.

To assemble the input for different dwrf writers, we compute
the index into the original input for every partition value
and wrap the original input into a DictionaryVector.
The DictionaryVector is sent to the corresponding dwrf writer
as writer input. In the frequent case of the input belonging
to the same partition, no input wrap happens and we directly
send the input vector to the dwrf writer.