Suppress errors for unused arguments of null-propagating functions
Null propagating functions will not be called if at least one argument
is null. Also, other arguments for rows where one argument was null
will not be evaluated for the null rows. This may cause errors to be
masked in a way depending on argument order. This makes a commutative
function like + not be commutative in the presence of masked errors.

We defer signalling of errors for arguments of null propagating
functions. An error will be signalled only if there is no null on its
row.

The errors for each Expr subtree are treated separately. Errors from
subtrees to the left must not null out rows for a subtree to the
right. Only arguments of the same function may interact in this way,
hence extra logic for scoping errors in Expr::evalAll.

Expr::evalAll works as before for functions with non-default null
behavior.  default null behavior, evalAll first captures the errors in
effect at entry. It then calls eval on each argument with an empty
error state in EvalCtx. It reads the errors from EvalCtx after each
argument and collects these in 'argumentErrors'. If an argument eval
returned null and no error for a row, subsequent arguments will not be
evaluated for this row. After

all arguments are evaluated, errors are cleared for those rows where
at least one argument produced a null and no error. If errors are left
at this time, they are added to the errors in effect at entry. If we
are in a mode where errors are signalled as they happen
(EvalCtx::throwOnError), we throw. Otherwise we remove the errors from
the selectivity vector for which we call the function. If the
selectivity vector goes empty, we return, otherwise we call the
function and produce a result or error for the surviving rows.